pyproject.toml
[project]
name = "music-lyric"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audio-separator[cpu]>=0.39.0",
    "beautifulsoup4>=4.14.2",
    "faster-whisper>=1.2.0",
    "mutagen>=1.47.0",
    "openai>=2.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
